# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EdeQR LMS API",
        "description": "bridge digital LMS projesi için REST API dokümantasyonu",
        "contact": {
            "email": "ahmet@aciksari.net"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/school/curriculum-overrides": {
            "get": {
                "tags": [
                    "Müfredat Yönetimi"
                ],
                "summary": "Tüm öğrenci müfredat geçersiz kılma kayıtlarını listeler.",
                "operationId": "indexStudentCurriculumOverride",
                "responses": {
                    "200": {
                        "description": "Başarılı işlem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StudentCurriculumOverride"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Yetkisiz İşlem"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Müfredat Yönetimi"
                ],
                "summary": "Yeni bir öğrenci müfredat geçersiz kılma kaydı oluşturur.",
                "operationId": "storeStudentCurriculumOverride",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreStudentCurriculumOverrideRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Başarılı oluşturma",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StudentCurriculumOverride"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Yetkisiz İşlem"
                    },
                    "422": {
                        "description": "Doğrulama Hatası"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/school/curriculum-overrides/{id}": {
            "get": {
                "tags": [
                    "Müfredat Yönetimi"
                ],
                "summary": "Belirli bir müfredat geçersiz kılma kaydını gösterir.",
                "operationId": "showStudentCurriculumOverride",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Geçersiz kılma kaydı ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı işlem",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StudentCurriculumOverride"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Yetkisiz İşlem"
                    },
                    "404": {
                        "description": "Kayıt bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Müfredat Yönetimi"
                ],
                "summary": "Belirli bir müfredat geçersiz kılma kaydını günceller.",
                "operationId": "updateStudentCurriculumOverride",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Geçersiz kılma kaydı ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateStudentCurriculumOverrideRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Başarılı güncelleme",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StudentCurriculumOverride"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Yetkisiz İşlem"
                    },
                    "404": {
                        "description": "Kayıt bulunamadı"
                    },
                    "422": {
                        "description": "Doğrulama Hatası"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Müfredat Yönetimi"
                ],
                "summary": "Belirli bir müfredat geçersiz kılma kaydını siler.",
                "operationId": "destroyStudentCurriculumOverride",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Geçersiz kılma kaydı ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Başarılı silme (İçerik Yok)"
                    },
                    "403": {
                        "description": "Yetkisiz İşlem"
                    },
                    "404": {
                        "description": "Kayıt bulunamadı"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreStudentCurriculumOverrideRequest": {
                "title": "Müfredat Geçersiz Kılma Oluşturma İsteği",
                "required": [
                    "student_id",
                    "curriculum_id",
                    "override_status"
                ],
                "properties": {
                    "student_id": {
                        "description": "Geçersiz kılınan öğrencinin ID'si",
                        "type": "integer",
                        "example": 105
                    },
                    "curriculum_id": {
                        "description": "Geçersiz kılınan Müfredat (Lesson/Unit) ID'si",
                        "type": "integer",
                        "example": 50
                    },
                    "override_status": {
                        "description": "Geçersiz kılma durumu (Assigned, Completed, Skipped)",
                        "type": "string",
                        "example": "Assigned"
                    },
                    "reason": {
                        "description": "Geçersiz kılma nedeni/açıklaması",
                        "type": "string",
                        "example": "Öğrenci bu üniteyi önceki okulunda tamamladı.",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateStudentCurriculumOverrideRequest": {
                "title": "Müfredat Geçersiz Kılma Güncelleme İsteği",
                "properties": {
                    "override_status": {
                        "description": "Geçersiz kılma durumu (Assigned, Completed, Skipped)",
                        "type": "string",
                        "example": "Completed"
                    },
                    "completion_date": {
                        "description": "Tamamlanma tarihi",
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-15",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "StudentCurriculumOverride": {
                "title": "StudentCurriculumOverride",
                "description": "Öğrenci müfredat geçiş (override) modeli",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "school_student_profile_id": {
                        "description": "Bağlı olduğu öğrenci profil ID",
                        "type": "integer",
                        "example": 12
                    },
                    "target_grade": {
                        "description": "Erişim verilecek sınıf seviyesi",
                        "type": "integer",
                        "example": 2
                    },
                    "unlimited_access": {
                        "description": "Tüm haftalara erişim izni",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-06T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-06T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT veya Sanctum tokenini \"Bearer {token}\" formatında girin"
            },
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "1": {
            "bearerAuth": []
        }
    }
}